backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
transformers==4.37.0
torch>=2.6.0
sentencepiece==0.1.99
protobuf==4.25.2
accelerate==0.26.0
langdetect==1.0.9
redis==5.0.1
celery==5.3.6
pymongo==4.6.1
motor==3.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.26.0
numpy>=1.26.3
sacremoses==0.1.1

